Add spring security: user accounts, authorization

+++ Recipes/task/src/recipes/service/RecipeService.java
@@ -1,21 +1,29 @@
 package recipes.service;
 
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import recipes.model.Recipe;
+import recipes.model.User;
 import recipes.repository.RecipeRepository;
+import recipes.repository.UserRepository;
 
+import java.security.AccessControlException;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class RecipeService {
     private final RecipeRepository recipeRepository;
+    private final UserRepository userRepository;
 
-    public RecipeService(RecipeRepository recipeRepository) {
+    public RecipeService(RecipeRepository recipeRepository, UserRepository userRepository) {
         this.recipeRepository = recipeRepository;
+        this.userRepository = userRepository;
     }
 
-    public long add(Recipe recipe) {
+    public long add(Recipe recipe, UserDetails user) {
+        recipe.setUser(userRepository.findUserByEmail(user.getUsername()).get());
         return recipeRepository.save(recipe).getId();
     }
 
@@ -31,8 +39,11 @@ public class RecipeService {
         return recipeRepository.findAllByNameContainingIgnoreCaseOrderByDateDesc(name);
     }
 
-    public void update(Long id, Recipe recipe) {
+    public void update(Long id, Recipe recipe, User author) {
         Recipe oldRecipe = recipeRepository.findById(id).orElseThrow();
+        if (!oldRecipe.getUser().equals(author)) {
+            throw new AccessControlException("Only the original author of the recipe is allowed to update it.");
+        }
         oldRecipe.setName(recipe.getName());
         oldRecipe.setCategory(recipe.getCategory());
         oldRecipe.setDescription(recipe.getDescription());
@@ -41,7 +52,12 @@ public class RecipeService {
         recipeRepository.save(oldRecipe);
     }
 
-    public void delete(Long id) {
+    public void delete(Long id, User author) {
+        Optional<Recipe> recipe = recipeRepository.findById(id);
+        recipe.orElseThrow(() -> new EmptyResultDataAccessException(1));
+        if (!recipe.get().getUser().equals(author)) {
+            throw new AccessControlException("Only the original author of the recipe is allowed to delete it.");
+        }
         recipeRepository.deleteById(id);
     }
 }
